Bug fixes.
Fixed a critical bug in the new rendering system.
Fixed a negative numbers related problem in the Identification system.
Fixed transparency problems for the Tab and the Tab-extended components.

+++ src/org/azentreprise/arionide/ui/overlay/components/Scroll.java
@@ -103,8 +103,8 @@ public class Scroll extends Tab {
 		super.updateAll();
 		
 		if(this.getBounds().getWidth() > 0) {
-			this.shadow = this.getBounds().getCenterX();
-			this.setShadowRadius(this.getBounds().getWidth() / 2);
+			this.shadow = (float) this.getBounds().getCenterX();
+			this.setShadowRadius((float) this.getBounds().getWidth() / 2);
 		}
 	}
